// src/app/ui/todoConsole/app.enum.ts
export enum MyApp {
    selectedOption = 'Selecciona una opcion',
    createTask = 'Create Task',
    listTask = 'List Task',
    nameTask = 'Name Task: ',
    descriptionTask = 'Description Task',
    enterContinue = 'Enter to continue',
    updateItem = 'Item to Update',
    deleteItem = 'Item to delete',
    updateTitle = 'Update Title - Yes or Not\n',
    updateDescription = 'Update Description - Yes or Not\n',
    checkYes = 'yes',
    checkNot = 'no',
    newTitle = 'New Title',
    newDescription = 'New Description',
    messageUpdate = 'Records updated successfully !!\n',
    messageDelete = 'Record Deleted Successfully'
}